create callback to get realEstateId when the entity gets inserted Inserting real estate entity and real estate photo entity both at the same time(function) Getting picUri and description by navigation library
Either now I have to disappear default real estate entity(id:0)
Or make the realEstateIdOfPhoto to nullable

+++ app/src/main/java/com/openclassrooms/realestatemanager/design_system/real_estate_photo/RealEstatePhotoItemViewState.kt
@@ -4,7 +4,7 @@ sealed class RealEstatePhotoItemViewState(
     val type: RealEstatePhotoType,
 ) {
     data class Content(
-        val photoId: Int,
+        val photoId: Long,
         val photoUrl: String,
         val photoDescription: String?,
     ):RealEstatePhotoItemViewState(RealEstatePhotoType.CONTENT)
